requirements.txt
Flask==3.1.3
psycopg2==2.9.11
python-dotenv==1.0.1
paramiko==4.0.0
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
PyJWT==2.12.1
flask-cors>=5.0.0
requests>=2.31.0
cryptography>=42.0.0